% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_info.R
\name{print_info}
\alias{print_info}
\title{Printing questions and labelled categories}
\usage{
print_info(data, like = "")
}
\arguments{
\item{data}{Data with labelled column types}
}
\value{

}
\description{
This function takes data from labelled column types and print
information from each column.
}
\examples{
print_info(mtcars)
}
